Refactor internal/mission package with Repository and Service interfaces
