+++ server/internal/app/route/mission.go
@@ -20,10 +20,10 @@ import (
 type MissionRoutes struct {
 	a agent.Service
 	j jwt.Service
-	m *mission.Service
+	m mission.Service
 }
 
-func NewMissionRoutes(a agent.Service, j jwt.Service, m *mission.Service) *MissionRoutes {
+func NewMissionRoutes(a agent.Service, j jwt.Service, m mission.Service) *MissionRoutes {
 	return &MissionRoutes{a: a, j: j, m: m}
 }
