+++ server/internal/mission/message.go
